api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
python-dotenv==1.0.1
PyJWT==2.9.0
passlib==1.7.4
python-multipart==0.0.9
openpyxl==3.1.5
redis==5.0.8
